fix quick timer button provider
